tar_preprocess.py
import os
import numpy as np
import pandas as pd
import tensorflow as tf
import scipy.stats as st
from PIL import Image
from scipy.misc import imread
from scipy.misc import imresize
from tensorflow.contrib.slim.nets import resnet_v1, inception, vgg

slim = tf.contrib.slim


# preprocess image before input the data

def preprocess_for_model(images, model_type):
    if 'inception' in model_type.lower():
        # images = tf.image.decode_jpeg(images)
        images = tf.image.convert_image_dtype(images, tf.float32)
        resize_shape = [1,224,224,3]
        images = tf.image.resize_images(images, size=[224,224], method=tf.image.ResizeMethod.AREA)
        # images = tf.image.resize_bilinear(images, [224,224],align_corners=False)
        # tensor-scalar operation
        images = (images / 255.0) * 2.0 - 1.0
        return images

    if 'resnet' in model_type.lower() or 'vgg' in model_type.lower():
        _R_MEAN = 123.68
        _G_MEAN = 116.78
        _B_MEAN = 103.94
        images = tf.image.resize_bilinear(images, [224,224],align_corners=False)
        tmp_0 = images[:,:,:,0] - _R_MEAN
        tmp_1 = images[:,:,:,1] - _G_MEAN
        tmp_2 = images[:,:,:,2] - _B_MEAN
        images = tf.stack([tmp_0,tmp_1,tmp_2],3)
        return images
