requirements.txt
absl-py==0.7.1
astor==0.7.1
gast==0.2.2
grpcio==1.20.0
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mock==2.0.0
numpy==1.16.3
pandas==0.24.2
pbr==5.1.3
Pillow==6.0.0
progressbar==2.5
protobuf==3.7.1
python-dateutil==2.8.0
pytz==2019.1
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.2
